# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SolidFireGateway, Version=1.1.1.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/SolidFire/{cluster}/volumes/foraccount": {
      "get": {
        "tags": [
          "Volumes"
        ],
        "parameters": [
          {
            "name": "cluster",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountID",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/SolidFire/{cluster}/volumes": {
      "post": {
        "tags": [
          "Volumes"
        ],
        "parameters": [
          {
            "name": "cluster",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVolumeParams"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVolumeParams"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVolumeParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreateVolumeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateVolumeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateVolumeResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Volumes"
        ],
        "parameters": [
          {
            "name": "cluster",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyVolumeParams"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyVolumeParams"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyVolumeParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ModifyVolumeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModifyVolumeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModifyVolumeResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Volumes"
        ],
        "parameters": [
          {
            "name": "cluster",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteVolumeParams"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteVolumeParams"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteVolumeParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/SolidFire/{cluster}/volumes/purge": {
      "post": {
        "tags": [
          "Volumes"
        ],
        "parameters": [
          {
            "name": "cluster",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurgeDeletedVolumeParams"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PurgeDeletedVolumeParams"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PurgeDeletedVolumeParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PurgeDeletedVolumeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurgeDeletedVolumeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurgeDeletedVolumeResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateVolumeParams": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "accountID": {
            "type": "integer",
            "format": "int32"
          },
          "totalSize": {
            "type": "integer",
            "format": "int64"
          },
          "enable512e": {
            "type": "boolean"
          },
          "attributes": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true
          },
          "qosPolicyID": {
            "type": "integer",
            "format": "int32"
          },
          "associateWithQoSPolicy": {
            "type": "boolean"
          },
          "qos": {
            "$ref": "#/components/schemas/Qos"
          }
        },
        "additionalProperties": false
      },
      "CreateVolumeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "result": {
            "$ref": "#/components/schemas/CreateVolumeResult"
          }
        },
        "additionalProperties": false
      },
      "CreateVolumeResult": {
        "type": "object",
        "properties": {
          "curve": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "volume": {
            "$ref": "#/components/schemas/Volume"
          },
          "volumeID": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "DeleteVolumeParams": {
        "type": "object",
        "properties": {
          "volumeID": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ModifyVolumeParams": {
        "type": "object",
        "properties": {
          "volumeID": {
            "type": "integer",
            "format": "int32"
          },
          "access": {
            "type": "string",
            "nullable": true
          },
          "attributes": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true
          },
          "qosPolicyID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "totalSize": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ModifyVolumeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "result": {
            "$ref": "#/components/schemas/ModifyVolumeResult"
          }
        },
        "additionalProperties": false
      },
      "ModifyVolumeResult": {
        "type": "object",
        "properties": {
          "volume": {
            "$ref": "#/components/schemas/Volume"
          }
        },
        "additionalProperties": false
      },
      "PurgeDeletedVolumeParams": {
        "type": "object",
        "properties": {
          "volumeID": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "PurgeDeletedVolumeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "result": {
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Qos": {
        "type": "object",
        "properties": {
          "burstIOPS": {
            "type": "integer",
            "format": "int32"
          },
          "burstTime": {
            "type": "integer",
            "format": "int32"
          },
          "curve": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "maxIOPS": {
            "type": "integer",
            "format": "int32"
          },
          "minIOPS": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}